new docs added and features

+++ book-inventory-service/src/main/kotlin/com/bookstore/bookinventory/dto/BookSearchRequestDTO.kt
@@ -1,12 +1,14 @@
 package com.bookstore.bookinventory.dto
 
+import java.math.BigDecimal
+
 data class BookSearchRequestDTO(
     val title: String? = null,
     val author: String? = null,
     val genre: String? = null,
     val isbn: String? = null,
-    val minPrice: Double? = null,
-    val maxPrice: Double? = null,
+    val minPrice: BigDecimal? = null,
+    val maxPrice: BigDecimal? = null,
     val genreFilter: String? = null,
     val availableOnly: Boolean? = null,
     val page: Int = 0,
